MenuHandler and Menu Logic
Changes made to MenuHandler for Console IO. Improved readability of code, prompts, and more. Many parts of the code are fully implemented but require testing. Some methods are not yet touched, they are marked with TODO.

Functionality Change: Rent will assume that the latest tenant of the apartment number will be used. Warnings are displayed to ensure correct documentation. There is currently no way to write to old Tenants (tenants who share the same apartment number but were input earlier). The warnings for adding Tenant and adding Rent make this clear.